requirements.txt
Django==5.2.7
djangorestframework
django-polymorphic
whitenoise
requests
beautifulsoup4
lxml
Pillow
pytz
packaging
channels
channels_redis
hypercorn
selenium